@@ -1,16 +1,10 @@
 # frozen_string_literal: true
 
 source "https://rubygems.org"
-
 ruby "2.7.1"
 
-git_source(:github) do |repo_name|
-  repo_name = "#{repo_name}/#{repo_name}" unless repo_name.include?("/")
-  "https://github.com/#{repo_name}.git"
-end
-
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem "rails", "~> 5.1.1"
+gem "rails", "~> 6.0.3"
 # Use postgresql as the database for Active Record
 gem "pg", "~> 0.18"
 
@@ -18,8 +12,6 @@ gem "aws-sdk", "~> 2"
 # bootstrap 4
 gem "bootstrap", "~> 4.3.1"
 gem "bstard"
-# Use CoffeeScript for .coffee assets and views
-gem "coffee-rails", "~> 4.2"
 gem "devise"
 gem "devise_invitable"
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
@@ -45,12 +37,11 @@ gem "whenever", require: false
 # Turbolinks makes navigating your web application faster. Read more: https://github.com/turbolinks/turbolinks
 gem "turbolinks", "~> 5"
 
-gem "passenger", "~> 5.1", require: false
-# Use Puma as the app server
-gem "puma", "~> 3.7"
-
 group :production do
   gem "airbrake", "~> 5.0"
+  # Use passenger as the app server in production. The environment web-ops have
+  # built currently expects this to be the case
+  gem "passenger", "~> 5.1", require: false
 end
 
 group :development do
@@ -61,15 +52,17 @@ group :development do
   gem "spring"
   gem "spring-watcher-listen", "~> 2.0.0"
   # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
-  gem "web-console", ">= 3.3.0"
+  gem "web-console"
 end
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem "byebug", platforms: %i[mri mingw x64_mingw]
+  gem "byebug"
   # Shim to load environment variables from a .env file into ENV in development
   # and test
   gem "dotenv-rails"
+  # Use Puma as the app server in development as test
+  gem "puma"
   # Project uses RSpec as its test framework
   gem "rspec-rails"
 end
@@ -82,7 +75,10 @@ group :test do
   gem "mocha"
   gem "rails-controller-testing"
   gem "selenium-webdriver"
-  gem "simplecov", require: false
+  # Generates a test coverage report on every `bundle exec rspec` call. We use
+  # the output to feed SonarCloud's stats and analysis. It does not yet support
+  # v0.18 hence locked to 0.17
+  gem "simplecov", "~> 0.17.1", require: false
   # Stubbing HTTP requests
   gem "webmock"
 end
@@ -94,6 +90,3 @@ group :benchmark do
   gem "rack-mini-profiler"
   gem "stackprof"
 end
-
-# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
-gem "tzinfo-data", platforms: %i[mingw mswin x64_mingw jruby]
