@@ -14,10 +14,10 @@ class ExcludeTransaction < ServiceObject
     if @transaction.updateable?
       old_user = Thread.current[:current_user]
       Thread.current[:current_user] = @user
-      @result = @transaction.update_attributes(excluded: true,
-                                               excluded_reason: @reason,
-                                               charge_calculation: nil,
-                                               tcm_charge: nil)
+      @result = @transaction.update(excluded: true,
+                                    excluded_reason: @reason,
+                                    charge_calculation: nil,
+                                    tcm_charge: nil)
 
       Thread.current[:current_user] = old_user
     else
