@@ -11,9 +11,9 @@ class UnapproveTransaction < ServiceObject
 
   def call
     @result = if @transaction.approved?
-                @transaction.update_attributes(approved_for_billing: false,
-                                               approver: @approver,
-                                               approved_for_billing_at: Time.zone.now)
+                @transaction.update(approved_for_billing: false,
+                                    approver: @approver,
+                                    approved_for_billing_at: Time.zone.now)
               else
                 false
               end
