@@ -59,7 +59,7 @@ class TransactionFileExporterTest < ActiveSupport::TestCase
   end
 
   def test_export_creates_file_with_all_approved_transactions
-    @transaction2.update_attributes(approved_for_billing: false)
+    @transaction2.update(approved_for_billing: false)
     @exporter.export
 
     file = TransactionFile.last
@@ -127,8 +127,8 @@ class TransactionFileExporterTest < ActiveSupport::TestCase
 
     file.transaction_details.update_all(customer_reference: "ABCD1234",
                                         reference_1: "ZZZZ9999")
-    @transaction2.update_attributes(tcm_charge: @transaction1.tcm_charge,
-                                    customer_reference: "AABBCCDD")
+    @transaction2.update(tcm_charge: @transaction1.tcm_charge,
+                         customer_reference: "AABBCCDD")
     # 1,2 and 3 same permit reference
     # 1 and 2 different customer reference
     # 3 different financial year
@@ -144,7 +144,7 @@ class TransactionFileExporterTest < ActiveSupport::TestCase
     file.transaction_details.update_all(customer_reference: "ABCD1234",
                                         reference_1: "ZZZZ9999",
                                         line_context_code: "E")
-    @transaction1.update_attributes(line_context_code: "A")
+    @transaction1.update(line_context_code: "A")
 
     # 1,2 and 3 same consent reference
     # 1 different context code
@@ -171,8 +171,8 @@ class TransactionFileExporterTest < ActiveSupport::TestCase
 
     file.transaction_details.update_all(customer_reference: "ABCD1234",
                                         reference_1: "ZZZZ9999")
-    @transaction2.update_attributes(tcm_charge: @transaction1.tcm_charge,
-                                    customer_reference: "AABBCCDD")
+    @transaction2.update(tcm_charge: @transaction1.tcm_charge,
+                         customer_reference: "AABBCCDD")
     # 1,2 and 3 same permit reference
     # 1 and 2 different customer reference
     # 3 different financial year
@@ -214,8 +214,8 @@ class TransactionFileExporterTest < ActiveSupport::TestCase
 
     file.transaction_details.update_all(customer_reference: "ABCD1234",
                                         reference_1: "ZZZZ9999")
-    @transaction2.update_attributes(tcm_charge: @transaction1.tcm_charge,
-                                    customer_reference: "AABBCCDD")
+    @transaction2.update(tcm_charge: @transaction1.tcm_charge,
+                         customer_reference: "AABBCCDD")
     # 1,2 and 3 same permit reference
     # 1 and 2 different customer reference
     # 3 different financial year
