@@ -4,10 +4,12 @@
 # The `threads` method setting takes two numbers: a minimum and maximum.
 # Any libraries that use thread pools should be configured to match
 # the maximum value specified for Puma. Default is set to 5 threads for minimum
-# and maximum; this matches the default thread size of Active Record.
+# and maximum; this matches the default thread size of Active Record. We set the
+# minimum to be 1 as currently we are only using puma in development and test
+# so typically we only expect to be running 1 thread.
 #
 threads_count = ENV.fetch("RAILS_MAX_THREADS", 5)
-threads threads_count, threads_count
+threads 1, threads_count
 
 # Specifies the `port` that Puma will listen on to receive requests; default is 3000.
 #
